+++ tessellation/procgen/rng_generator.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 import numpy as np
 
-from tessellation.procgen.generator import Action, Generator
+from tessellation.procgen.generator import Action, Generator, GenerationResult
 
 VALID_ACTIONS = [
     Action.UP,
@@ -23,27 +23,34 @@ class RNGGenerator(Generator):
         self,
         side_len: int,
         action_probs: Optional[list[Action]] = None,
-    ) -> np.ndarray:
+    ) -> GenerationResult:
         """Generate a tesselation mask using a random number generator."""
         if action_probs is None:
             action_probs = [1 / len(VALID_ACTIONS)] * len(VALID_ACTIONS)
 
         assert len(VALID_ACTIONS) == len(action_probs)
 
-        y_mask = self._generate_side(side_len, action_probs)
-        x_mask = self._generate_side(side_len, action_probs).T
-        return x_mask | y_mask
+        y_axis_mask, y_line_actions = self._generate_side(side_len, action_probs)
+        x_axis_mask, x_line_actions = self._generate_side(side_len, action_probs)
+        # Transpose the x_mask to make the line run along the x-axis
+        x_axis_mask = x_axis_mask.T
+        return GenerationResult(
+            mask=x_axis_mask | y_axis_mask,
+            line_actions=[y_line_actions, x_line_actions],
+        )
 
-    def _generate_side(self, side_len: int, action_probs: list) -> np.ndarray:
+    def _generate_side(
+        self, side_len: int, action_probs: list
+    ) -> tuple[np.ndarray, list[Action]]:
         cursor = {"x": 0, "y": 0}
         action_list = []
         while cursor["x"] < side_len - 1:
+
             action = self._get_rand_action(
                 actions_list=VALID_ACTIONS, action_probs=action_probs
             )
             action_list.append(action)
-            # Note: it's probably inefficient to call flood_fill on every step, but for
-            # now, it's giving the most consistent results.
+
             if action == Action.UP:
                 cursor["y"] -= 1
             elif action == Action.UP_RIGHT:
@@ -60,7 +67,7 @@ class RNGGenerator(Generator):
                 raise ValueError(f"Unsupported action: {action}")
 
         mask = np.zeros((side_len, side_len), dtype=int)
-        return self._draw_line(mask, (0, 0), action_list)
+        return self._draw_line(mask, (0, 0), action_list), action_list
 
     def _get_rand_action(
         self, actions_list: list[Action], action_probs: list[float]