+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Track the changes that have been made to the project over time.
 
 ## Version 0.0
 
+### 0.0.3 (****-**-**)
+- Add simple genetic algorithm code for generating tessellation
+- Refactor generation code to have line drawing in generator base class
+
 ### 0.0.2 (2024-08-15)
 
 - Added RNG tessellation generation code and drawing code